# openapi.yaml
openapi: 3.0.0
info:
  title: Second Brain Server API
  description: |
    Comprehensive personal intelligence platform API for knowledge management, creativity, learning, and personal growth.

    ## Features
    - **Stateless Authentication**: JWT-based authentication with no server-side sessions
    - **OAuth Integration**: Google OAuth support with CSRF protection
    - **Database Management**: Create and manage custom databases with flexible schemas
    - **Record Management**: CRUD operations for database records with rich property types
    - **User Management**: Complete user lifecycle management with role-based access control
    - **Rate Limiting**: Built-in rate limiting for security and performance
    - **File Operations**: Import/export data in multiple formats (JSON, CSV, Excel)
    - **Real-time Collaboration**: Share databases with granular permissions

    ## Authentication
    This API uses JWT Bearer tokens for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Rate Limiting
    - Login: 5 attempts per 15 minutes
    - Registration: 3 attempts per hour
    - Password Reset: 3 attempts per hour
    - Token Refresh: 10 attempts per 15 minutes
    - OAuth: 20 attempts per 15 minutes
    - General API: 100 requests per 15 minutes

    ## Database Property Types
    - **TEXT**: Single-line text
    - **NUMBER**: Numeric values
...
    - **DATE**: Date and time values
...
  version: 1.0.0
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation successful
        data:
          type: object
          description: Response data (varies by endpoint)
        error:
          type: object
          properties:
            message:
              type: string
              example: Error description
            statusCode:
              type: integer
              example: 400
            status:
              type: string
              example: error
    ValidationError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
              example: Request body validation failed
            statusCode:
              type: integer
              example: 400
            status:
              type: string
              example: error
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: email
                  message:
                    type: string
                    example: Please provide a valid email address
                  value:
                    type: string
                    example: invalid-email
...
    Database:
      type: object
      properties:
        id:
          type: string
          description: Unique database identifier
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          description: Database name
          example: Project Tasks
        description:
          type: string
          description: Database description
          example: Track project tasks and their progress
        icon:
          type: string
          description: Database icon
          example: 📋
        cover:
          type: string
          format: uri
          description: Database cover image URL
          example: https://example.com/cover.jpg
        workspaceId:
          type: string
          description: Parent workspace ID
          example: 507f1f77bcf86cd799439012
        ownerId:
          type: string
          description: Database owner user ID
          example: 507f1f77bcf86cd799439013
        isPublic:
          type: boolean
          description: Whether database is publicly accessible
          example: false
        isFavorite:
          type: boolean
          description: Whether database is marked as favorite
          example: false
        categoryId:
          type: string
          description: Category ID this database belongs to
          example: 507f1f77bcf86cd799439020
        tags:
          type: array
          items:
            type: string
          description: Database tags for organization
          example: ["work", "project", "urgent"]
        lastAccessedAt:
          type: string
          format: date-time
          description: When database was last accessed
          example: 2023-01-01T10:30:00Z
        accessCount:
          type: integer
          description: Number of times database has been accessed
          example: 42
        properties:
          type: array
          items:
            $ref: '#/components/schemas/DatabaseProperty'
        views:
          type: array
          items:
            $ref: '#/components/schemas/DatabaseView'
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/DatabasePermission'
        createdAt:
          type: string
          format: date-time
          example: 2023-01-01T00:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2023-01-01T12:00:00Z
...
    DatabaseProperty:
      type: object
      properties:
        id:
          type: string
          description: Property identifier
          example: prop_001
        name:
          type: string
          description: Property name
          example: Task Name
        type:
          type: string
          enum: [TEXT, NUMBER, SELECT, MULTI_SELECT, DATE, CHECKBOX, URL, EMAIL, PHONE, RELATION, FORMULA, ROLLUP, CREATED_TIME, CREATED_BY, LAST_EDITED_TIME, LAST_EDITED_BY]
          description: Property type
          example: TEXT
        description:
          type: string
          description: Property description
          example: The name of the task
        required:
          type: boolean
          description: Whether property is required
          example: true
        isVisible:
          type: boolean
          description: Whether property is visible in views
          example: true
        order:
          type: integer
          description: Property display order
          example: 1
        selectOptions:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                example: opt_001
              name:
                type: string
                example: High Priority
              color:
                type: string
                example: "#FF0000"
        relationConfig:
          type: object
          properties:
            relatedDatabaseId:
              type: string
              example: 507f1f77bcf86cd799439014
            relationType:
              type: string
              enum: [ONE_TO_ONE, ONE_TO_MANY, MANY_TO_ONE, MANY_TO_MANY]
              example: MANY_TO_ONE
        formulaConfig:
          type: object
          properties:
            expression:
              type: string
              example: prop("Priority") + " - " + prop("Status")
            returnType:
              type: string
              enum: [TEXT, NUMBER, DATE, CHECKBOX]
              example: TEXT
...
    DatabaseView:
      type: object
      properties:
        id:
          type: string
          description: View identifier
          example: view_001
        name:
          type: string
          description: View name
          example: All Tasks
        type:
          type: string
          enum: [TABLE, BOARD, TIMELINE, CALENDAR, GALLERY, LIST]
          description: View type
          example: TABLE
        isDefault:
          type: boolean
          description: Whether this is the default view
          example: true
        filters:
          type: array
          items:
            type: object
            required: [propertyId, operator]
            properties:
              propertyId:
                type: string
                example: prop_001
              operator:
                type: string
                example: contains
              value:
                description: Filter value (type varies)
                example: urgent
        sorts:
          type: array
          items:
            type: object
            required: [propertyId, direction]
            properties:
              propertyId:
                type: string
                example: prop_002
              direction:
                type: string
                enum: [asc, desc]
                example: desc
        groupBy:
          type: string
          description: Property ID to group by
          example: prop_003
        visibleProperties:
          type: array
          items:
            type: string
          description: List of visible property IDs
          example: ["prop_001", "prop_002", "prop_003"]
...
    DatabaseRecord:
      type: object
      properties:
        id:
          type: string
          description: Record identifier
          example: 507f1f77bcf86cd799439015
        databaseId:
          type: string
          description: Parent database ID
          example: 507f1f77bcf86cd799439011
        properties:
          type: object
          description: Record property values
          example:
            prop_001: "Complete user authentication"
            prop_002: "High"
            prop_003: true
        createdAt:
          type: string
          format: date-time
          example: 2023-01-01T00:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2023-01-01T12:00:00Z
        createdBy:
          type: string
          description: User ID who created the record
          example: 507f1f77bcf86cd799439013
...
    DatabasePermission:
      type: object
      properties:
        userId:
          type: string
          description: User ID
          example: 507f1f77bcf86cd799439013
        permission:
          type: string
          enum: [read, write, admin]
          description: Permission level
          example: write
        grantedAt:
          type: string
          format: date-time
          example: 2023-01-01T00:00:00Z

    # Database Category Schemas
...
paths:
...
  /search:
    get:
      tags: [Search & Analytics]
      summary: Global search
      description: |
        Search across all content types including databases, records, and files.

        **Features**:
        - Cross-content search
        - Type-specific filtering
        - Relevance scoring
        - Pagination support
      operationId: globalSearch
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
            minLength: 1
        - name: type
          in: query
          description: Content type filter
          schema:
            type: string
            enum: [all, databases, records, files]
            default: all
        - name: limit
          in: query
          description: Maximum results to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Search completed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          databases:
                            type: array
                            items:
                              $ref: '#/components/schemas/Database'
                          records:
                            type: array
                            items:
                              $ref: '#/components/schemas/DatabaseRecord'
                          files:
                            type: array
                            items:
                              $ref: '#/components/schemas/FileResponse'
                          total:
                            type: integer
                            description: Total number of results
        '400':
          description: Invalid search query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /search/databases:
    get:
      tags: [Search & Analytics]
      summary: Search databases
      description: |
        Search specifically within databases by name, description, or properties.

        **Features**:
        - Database-specific search
        - Metadata search
        - Access control filtering
      operationId: searchDatabases
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
            minLength: 1
        - name: limit
          in: query
          description: Maximum results to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Database search completed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          databases:
                            type: array
                            items:
                              $ref: '#/components/schemas/Database'
                          total:
                            type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /search/records:
    get:
      tags: [Search & Analytics]
      summary: Search records
      description: |
        Search within database records by content and properties.

        **Features**:
        - Record content search
        - Property-based filtering
        - Database-specific search
        - Full-text search capabilities
      operationId: searchRecords
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
            minLength: 1
        - name: databaseId
          in: query
          description: Limit search to specific database
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum results to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Record search completed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          records:
                            type: array
                            items:
                              $ref: '#/components/schemas/DatabaseRecord'
                          total:
                            type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /search/suggestions:
    get:
      tags: [Search & Analytics]
      summary: Get search suggestions
      description: |
        Get search suggestions based on user's query and content.

        **Features**:
        - Auto-complete suggestions
        - Recent searches
        - Popular searches
        - Context-aware suggestions
      operationId: getSearchSuggestions
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          description: Partial search query
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: Search suggestions retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            suggestion:
                              type: string
                              description: Suggested search term
                            type:
                              type: string
                              enum: [database, record, file, recent, popular]
                              description: Type of suggestion
                            count:
                              type: integer
                              description: Number of results for this suggestion
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /tags:
    get:
      tags: [Tags Management]
      summary: Get user's tags
      description: |
        Retrieve all tags created by the user.

        **Features**:
        - User-specific tags
        - Search and filtering
        - Sorting options
        - Usage statistics
      operationId: getUserTags
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          description: Search tags by name
          schema:
            type: string
        - name: sortBy
          in: query
          description: Sort field
          schema:
            type: string
            enum: [name, createdAt, updatedAt]
            default: name
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: Tags retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Tag'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    post:
      tags: [Tags Management]
      summary: Create new tag
      description: |
        Create a new tag for organizing content.

        **Features**:
        - Custom tag names
        - Color coding
        - Optional descriptions
        - Duplicate prevention
      operationId: createTag
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 50
                  description: Tag name
                  example: Important
                color:
                  type: string
                  pattern: '^#[0-9A-F]{6}$'
                  description: Hex color code
                  example: '#FF5733'
                description:
                  type: string
                  maxLength: 200
                  description: Optional tag description
                  example: High priority items
              required:
                - name
      responses:
        '201':
          description: Tag created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Tag'
        '400':
          description: Validation error or duplicate tag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /tags/{id}:
    get:
      tags: [Tags Management]
      summary: Get tag by ID
      description: |
        Retrieve detailed information about a specific tag.

        **Features**:
        - Tag details and metadata
        - Usage statistics
        - Associated content count
      operationId: getTagById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Tag ID
          schema:
            type: string
      responses:
        '200':
          description: Tag retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Tag'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Tag not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    put:
      tags: [Tags Management]
      summary: Update tag
      description: |
        Update an existing tag's properties.

        **Features**:
        - Update name, color, description
        - Validation and duplicate checking
        - Preserve tag associations
      operationId: updateTag
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Tag ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 50
                  description: Tag name
                color:
                  type: string
                  pattern: '^#[0-9A-F]{6}$'
                  description: Hex color code
                description:
                  type: string
                  maxLength: 200
                  description: Tag description
      responses:
        '200':
          description: Tag updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Tag'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Tag not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    delete:
      tags: [Tags Management]
      summary: Delete tag
      description: |
        Delete a tag permanently.

        **Features**:
        - Permanent tag deletion
        - Remove tag associations
        - Cleanup related data
      operationId: deleteTag
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Tag ID
          schema:
            type: string
      responses:
        '200':
          description: Tag deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Tag not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # Notifications Management Endpoints
...