# src/modules/search/controllers/search.controller.ts
import { Request, Response, NextFunction } from 'express';
import { catchAsync, sendSuccessResponse, createNotFoundError } from '../../../utils';
import { AuthenticatedRequest } from '../../../middlewares/auth';

/**
 * Global search across all content
 */
export const globalSearch = catchAsync(
  async (req: Request, res: Response, next: NextFunction): Promise<void> => {
    const userId = (req as AuthenticatedRequest).user.userId;
    if (!userId) return next(createNotFoundError('User authentication required'));

    const { q, type, limit = 20 } = req.query;
    
    // TODO: Implement global search logic
    const results = {
      databases: [],
      records: [],
      files: [],
      total: 0
    };

    sendSuccessResponse(res, results, 'Search completed successfully');
  }
);

/**
 * Search databases
 */
export const searchDatabases = catchAsync(
  async (req: Request, res: Response, next: NextFunction): Promise<void> => {
    const userId = (req as AuthenticatedRequest).user.userId;
    if (!userId) return next(createNotFoundError('User authentication required'));

    const { q, limit = 20 } = req.query;
    
    // TODO: Implement database search logic
    const results = {
      databases: [],
      total: 0
    };

    sendSuccessResponse(res, results, 'Database search completed successfully');
  }
);

/**
 * Search records
 */
export const searchRecords = catchAsync(
  async (req: Request, res: Response, next: NextFunction): Promise<void> => {
    const userId = (req as AuthenticatedRequest).user.userId;
    if (!userId) return next(createNotFoundError('User authentication required'));

    const { q, databaseId, limit = 20 } = req.query;
    
    // TODO: Implement record search logic
    const results = {
      records: [],
      total: 0
    };

    sendSuccessResponse(res, results, 'Record search completed successfully');
  }
);

/**
 * Get search suggestions
 */
export const getSearchSuggestions = catchAsync(
  async (req: Request, res: Response, next: NextFunction): Promise<void> => {
    const userId = (req as AuthenticatedRequest).user.userId;
    if (!userId) return next(createNotFoundError('User authentication required'));

    const { q } = req.query;
    
    // TODO: Implement search suggestions logic
    const suggestions = [];

    sendSuccessResponse(res, suggestions, 'Search suggestions retrieved successfully');
  }
);