# src/modules/tags/controllers/tags.controller.ts
import { Request, Response, NextFunction } from 'express';
import { catchAsync, sendSuccessResponse, createNotFoundError } from '../../../utils';
import { AuthenticatedRequest } from '../../../middlewares/auth';

/**
 * Get user's tags
 */
export const getUserTags = catchAsync(
  async (req: Request, res: Response, next: NextFunction): Promise<void> => {
    const userId = (req as AuthenticatedRequest).user.userId;
    if (!userId) return next(createNotFoundError('User authentication required'));

    const { search, sortBy = 'name', sortOrder = 'asc' } = req.query;
    
    // TODO: Implement get user tags logic
    const tags = [];

    sendSuccessResponse(res, tags, 'Tags retrieved successfully');
  }
);

/**
 * Create new tag
 */
export const createTag = catchAsync(
  async (req: Request, res: Response, next: NextFunction): Promise<void> => {
    const userId = (req as AuthenticatedRequest).user.userId;
    if (!userId) return next(createNotFoundError('User authentication required'));

    const { name, color, description } = req.body;
    
    // TODO: Implement create tag logic
    const tag = {
      id: 'temp-id',
      name,
      color,
      description,
      userId,
      createdAt: new Date(),
      updatedAt: new Date()
    };

    sendSuccessResponse(res, tag, 'Tag created successfully', 201);
  }
);

/**
 * Get tag by ID
 */
export const getTagById = catchAsync(
  async (req: Request, res: Response, next: NextFunction): Promise<void> => {
    const userId = (req as AuthenticatedRequest).user.userId;
    if (!userId) return next(createNotFoundError('User authentication required'));

    const { id } = req.params;
    
    // TODO: Implement get tag by ID logic
    const tag = null;

    if (!tag) {
      return next(createNotFoundError('Tag not found'));
    }

    sendSuccessResponse(res, tag, 'Tag retrieved successfully');
  }
);

/**
 * Update tag
 */
export const updateTag = catchAsync(
  async (req: Request, res: Response, next: NextFunction): Promise<void> => {
    const userId = (req as AuthenticatedRequest).user.userId;
    if (!userId) return next(createNotFoundError('User authentication required'));

    const { id } = req.params;
    const { name, color, description } = req.body;
    
    // TODO: Implement update tag logic
    const tag = null;

    if (!tag) {
      return next(createNotFoundError('Tag not found'));
    }

    sendSuccessResponse(res, tag, 'Tag updated successfully');
  }
);

/**
 * Delete tag
 */
export const deleteTag = catchAsync(
  async (req: Request, res: Response, next: NextFunction): Promise<void> => {
    const userId = (req as AuthenticatedRequest).user.userId;
    if (!userId) return next(createNotFoundError('User authentication required'));

    const { id } = req.params;
    
    // TODO: Implement delete tag logic

    sendSuccessResponse(res, null, 'Tag deleted successfully');
  }
);